# docs/swagger.yaml
basePath: /api/v1
definitions:
  features.HTTPAuthError:
    properties:
      error:
        example: Mensagem de erro aqui
        type: string
    type: object
  features.SignInBodyDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  features.SignUpBodyDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API para gerenciamento de produtos.
  title: API de Desafio Go
  version: "1.0"
paths:
...
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais de login
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/features.SignInBodyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/features.HTTPAuthError'
      summary: Realiza o login
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados de registro
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/features.SignUpBodyDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registra um novo usuário
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"